refactor: exclude non-route GameBanana files from route scanning
