+++ src/renderer/src/routes/gamebanana/-cards/mod-grid-card.tsx
@@ -1,4 +1,4 @@
-import type { SubmissionListItem } from "../types";
+import type { SubmissionListItem } from "../-types";
 import { PreviewCard } from "./preview-card";
 
 export function ModGridCard({
