+++ src/renderer/src/routes/gamebanana/-shared/common.tsx
