+++ src/renderer/src/routes/gamebanana/-types.ts
